+++ barracuda_secure_edge/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 <!-- towncrier release notes start -->
 
-## 1.0.0 / 2025-10-30
+## 1.1.0 / 2025-11-26
+
+***Added***:
+
+* Bump Python to 3.13 ([#21848](https://github.com/DataDog/integrations-core/pull/21848))
+* Bump minimum version of datadog-checks-base to 37.24.0 ([#21945](https://github.com/DataDog/integrations-core/pull/21945))
+
+## 1.0.0 / 2025-10-30 / Agent 7.73.0
 
 ***Added***: